fix, yq version 3.3.2 was detected as unexpected
when using yq version 3.3.2, it was incorrectly outputting an error:

$ ./toolchain-to-template.sh
https://cloud.ibm.com/devops/toolchains/32ae1415-e0a8-45ee-9b0e-84596e3709ae?env_id=ibm:yp:us-south
yq version 3.3.2
Unexpected prereq 'yq --version' is not 2.x or 3.x
$

as it considered that version not to be a valid 3.x version.
Problem was that yq --version was outputting to stderr
instead of stdout.
Fix is to use yq --version output from either stderr or stdout